Update tinyMCE marvinjs plugin [SCI-7154]
